Break out a function to avoid code duplication
