+++ src/functions/operatorname.js
@@ -3,6 +3,7 @@ import defineFunction, {ordargument} from "../defineFunction";
 import buildCommon from "../buildCommon";
 import mathMLTree from "../mathMLTree";
 import {SymbolNode} from "../domTree";
+import {assembleSupSub} from "./assembleSupSub";
 import {assertNodeType, checkNodeType} from "../parseNode";
 
 import * as html from "../buildHTML";
@@ -68,90 +69,9 @@ export const htmlBuilder: HtmlBuilderSupSub<"operatorname"> = (grp, options) => 
     }
 
     if (hasLimits) {
-        let sub;
-        let sup;
-        // We manually have to handle the superscripts and subscripts. This,
-        // aside from the kern calculations, is copied from supsub.
-        if (supGroup) {
-            const elem = html.buildGroup(
-                supGroup, options.havingStyle(options.style.sup()), options);
+        return assembleSupSub(base, supGroup, subGroup, options,
+            options.style, 0, 0);
 
-            sup = {
-                elem,
-                kern: Math.max(
-                    options.fontMetrics().bigOpSpacing1,
-                    options.fontMetrics().bigOpSpacing3 - elem.depth),
-            };
-        }
-
-        if (subGroup) {
-            const elem = html.buildGroup(
-                subGroup, options.havingStyle(options.style.sub()), options);
-
-            sub = {
-                elem,
-                kern: Math.max(
-                    options.fontMetrics().bigOpSpacing2,
-                    options.fontMetrics().bigOpSpacing4 - elem.height),
-            };
-        }
-
-        // Build the final group as a vlist of the possible subscript, base,
-        // and possible superscript.
-        let finalGroup;
-        if (sup && sub) {
-            const bottom = options.fontMetrics().bigOpSpacing5 +
-                sub.elem.height + sub.elem.depth +
-                sub.kern +
-                base.depth;
-
-            finalGroup = buildCommon.makeVList({
-                positionType: "bottom",
-                positionData: bottom,
-                children: [
-                    {type: "kern", size: options.fontMetrics().bigOpSpacing5},
-                    {type: "elem", elem: sub.elem},
-                    {type: "kern", size: sub.kern},
-                    {type: "elem", elem: base},
-                    {type: "kern", size: sup.kern},
-                    {type: "elem", elem: sup.elem},
-                    {type: "kern", size: options.fontMetrics().bigOpSpacing5},
-                ],
-            }, options);
-        } else if (sub) {
-            const top = base.height;
-            finalGroup = buildCommon.makeVList({
-                positionType: "top",
-                positionData: top,
-                children: [
-                    {type: "kern", size: options.fontMetrics().bigOpSpacing5},
-                    {type: "elem", elem: sub.elem},
-                    {type: "kern", size: sub.kern},
-                    {type: "elem", elem: base},
-                ],
-            }, options);
-        } else if (sup) {
-            const bottom = base.depth;
-
-            finalGroup = buildCommon.makeVList({
-                positionType: "bottom",
-                positionData: bottom,
-                children: [
-                    {type: "elem", elem: base},
-                    {type: "kern", size: sup.kern},
-                    {type: "elem", elem: sup.elem},
-                    {type: "kern", size: options.fontMetrics().bigOpSpacing5},
-                ],
-            }, options);
-        } else {
-            // This case probably shouldn't occur (this would mean the
-            // supsub was sending us a group with no superscript or
-            // subscript) but be safe.
-            return base;
-        }
-
-        return buildCommon.makeSpan(
-            ["mop", "op-limits"], [finalGroup], options);
     } else {
         return base;
     }
